requirements.txt
Flask==2.0.1
tensorflow~=2.6.0
pillow~=8.3.1
numpy~=1.19.5
requests~=2.26.0
matplotlib>=3.2.2
numpy>=1.18.5
opencv-python>=4.1.2
Pillow
PyYAML>=5.3.1
scipy>=1.4.1
torch>=1.7.0
torchvision>=0.8.1
tqdm>=4.41.0
seaborn>=0.11.0
pandas
webcolors~=1.11.1